refactor: separate out hamburger component from menu items
